requirements.txt
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
future==0.18.2
numpy==1.17.4
opencv-python==4.1.2.30
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
